联调基本完成，待测试datax执行效果

+++ arrange/src/main/java/com/cyj/arrange/service/TaskService.java
@@ -3,6 +3,8 @@ package com.cyj.arrange.service;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.cyj.arrange.bean.vo.TCfgPipelineTaskVO;
+import com.cyj.arrange.bean.vo.TCfgPipelineVO;
 import com.cyj.arrange.entry.TCfgPipeline;
 import com.cyj.arrange.entry.TCfgPipelineTask;
 import com.cyj.arrange.entry.TCfgTask;
@@ -10,6 +12,7 @@ import com.cyj.arrange.mapper.TCfgPipelineMapper;
 import com.cyj.arrange.mapper.TCfgPipelineTaskMapper;
 import com.cyj.arrange.mapper.TCfgTaskMapper;
 import org.apache.commons.lang.StringUtils;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -47,19 +50,27 @@ public class TaskService {
         }
     }
 
-    public void delTask(Integer taskID)
+    public boolean delTask(Integer taskID)
     {
+        boolean rs = false;
         if (taskID!=null)
         {
-            tCfgTaskMapper.deleteById(taskID);
+            if (tCfgPipelineTaskMapper.selectCount(new QueryWrapper<TCfgPipelineTask>().eq("task_id",taskID))==0)
+            {
+                tCfgTaskMapper.deleteById(taskID);
+                rs = true;
+            }
         }
+        return rs;
     }
 
+    @Transactional
     public void delPipeline(Integer pipelineID)
     {
         if (pipelineID!=null)
         {
             tCfgPipelineMapper.deleteById(pipelineID);
+            tCfgPipelineTaskMapper.delete(new QueryWrapper<TCfgPipelineTask>().eq("pipeline_id",pipelineID));
         }
     }
 
@@ -77,16 +88,14 @@ public class TaskService {
         }
     }
 
-    @Transactional
-    public void setPipelineTask(Integer pipelineID, List<Tuple2<Integer,Integer>> tasks)
+    public void addPipelineTask(TCfgPipelineTask pipelineTask)
     {
-        tCfgPipelineTaskMapper.delete(new QueryWrapper<TCfgPipelineTask>().eq("pipeline_id",pipelineID));
-        for (Tuple2<Integer,Integer> tuple:tasks)
-        {
-            TCfgPipelineTask tCiPipelineTask = new TCfgPipelineTask();
-            tCiPipelineTask.setPipelineId(pipelineID).setTaskId(tuple.getT1()).setOrderNo(tuple.getT2());
-            tCfgPipelineTaskMapper.insert(tCiPipelineTask);
-        }
+        tCfgPipelineTaskMapper.insert(pipelineTask);
+    }
+
+    public void delPipelineTask(Integer seqId)
+    {
+        tCfgPipelineTaskMapper.deleteById(seqId);
     }
 
     public IPage<TCfgTask> queryTaskPager(String taskName,Integer ownerID,int pageNo,int pageSize)
@@ -104,14 +113,23 @@ public class TaskService {
     public IPage<TCfgPipeline> queryPipelinePager(String pipelineName,Integer ownerID,int pageNo,int pageSize)
     {
         QueryWrapper<TCfgPipeline> queryWrapper = new QueryWrapper<>();
-        queryWrapper.like("name",pipelineName).eq("owner",ownerID);
+        if (StringUtils.isNotEmpty(pipelineName))
+        {
+            queryWrapper.like("name",pipelineName);
+        }
+        queryWrapper.eq("owner",ownerID);
         Page<TCfgPipeline> page = new Page<>(pageNo,pageSize);
         return tCfgPipelineMapper.selectPage(page,queryWrapper);
     }
 
-    public List<TCfgPipelineTask> queryPipelineTask(Integer pipelineID)
+    public TCfgPipelineVO queryPipeline(Integer pipelineID)
     {
-        return tCfgPipelineTaskMapper.selectList(new QueryWrapper<TCfgPipelineTask>().eq("pipeline_id",pipelineID));
+        TCfgPipeline pipeline = tCfgPipelineMapper.selectById(pipelineID);
+        List<TCfgPipelineTaskVO> tasks = tCfgTaskMapper.findTaskByPipelineID(pipelineID);
+        TCfgPipelineVO vo = new TCfgPipelineVO();
+        BeanUtils.copyProperties(pipeline,vo);
+        vo.setTasks(tasks);
+        return vo;
     }
 
     public List<TCfgTask> queryAllTask(Integer userID)
